@@ -1,28 +1,28 @@
 group 'id.kakzaki.blue_thermal_printer'
-version '1.1.3'
+
 
 buildscript {
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.0'
+        classpath 'com.android.tools.build:gradle:3.5.4'
     }
 }
 
 rootProject.allprojects {
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
 }
 
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
@@ -34,10 +34,9 @@ android {
     lintOptions {
         disable 'InvalidPackage'
     }
-    buildToolsVersion '29.0.2'
 }
 dependencies {
-    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.appcompat:appcompat:1.4.2'
     implementation 'com.google.zxing:core:3.3.3'
     implementation 'com.android.support:multidex:1.0.3'
     implementation 'com.journeyapps:zxing-android-embedded:3.6.0@aar'
